+++ apps/menu-bar/metro.config.js
@@ -21,17 +21,6 @@ config.resolver.sourceExts = [
   'svg', // react-native-svg-transformer
 ];
 
-config.resolver.resolveRequest = (context, moduleName, platform) => {
-  if (
-    platform === 'macos' &&
-    (moduleName === 'react-native' || moduleName.startsWith('react-native/'))
-  ) {
-    const newModuleName = moduleName.replace('react-native', 'react-native-macos');
-    return context.resolveRequest(context, newModuleName, platform);
-  }
-  return context.resolveRequest(context, moduleName, platform);
-};
-
 config.transformer.babelTransformerPath = require.resolve('react-native-svg-transformer/expo');
 config.transformer.getTransformOptions = async () => ({
   transform: {
@@ -39,16 +28,5 @@ config.transformer.getTransformOptions = async () => ({
     inlineRequires: true,
   },
 });
-
-const originalGetModulesRunBeforeMainModule = config.serializer.getModulesRunBeforeMainModule;
-config.serializer.getModulesRunBeforeMainModule = () => {
-  try {
-    return [
-      require.resolve('react-native/Libraries/Core/InitializeCore'),
-      require.resolve('react-native-macos/Libraries/Core/InitializeCore'),
-    ];
-  } catch {}
-  return originalGetModulesRunBeforeMainModule();
-};
 
 module.exports = config;